dom/media/webcodecs/test/test_videoColorSpace.html

<!DOCTYPE HTML>
<html>
<head>
  <title></title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
let defaultVideoColorSpaceInit = {
  primaries: "bt709",
  transfer: "bt709",
  matrix: "bt709",
  fullRange: true,
};
let VCSIs = generateAllCombinations(defaultVideoColorSpaceInit);
for (let VCSI of VCSIs) {
  check(VCSI);
}

function generateAllCombinations(obj) {
  let keySets = getAllSubsets(Object.keys(obj));
  let combinations = [];
  for (let set of keySets) {
    let element = {};
    for (let key of set) {
      element[key] = obj[key];
    }
    combinations.push(element);
  }
  return combinations;
}

function getAllSubsets(set) {
  let sebsets = [];
  let upperbound = 1 << set.length;
  for (let bitmask = 0; bitmask < upperbound; ++bitmask) {
    let subset = [];
    for (let i = 0; i < set.length; ++i) {
      if (bitmask & (1 << i)) {
        subset.push(set[i]);
      }
    }
    sebsets.push(subset);
  }
  return sebsets;
}

function check(VCSI) {
  let vcs = new VideoColorSpace(VCSI);
  let json = vcs.toJSON();
  const VCSIKeys = Object.keys(VCSI);
  is(
    Object.keys(json).length,
    VCSIKeys.length,
    "VideoColorSpace.ToJSON() should have same number of properties as the given VideoColorSpaceInit's one"
  );
  for (const p of VCSIKeys) {
    is(
      json[p],
      VCSI[p],
      "VideoColorSpace.ToJSON()'s property should match the given VideoColorSpaceInit's one"
    );
  }
}
</script>
</body>
</html>
